did a few changes to optimize performance
on my laptop (when its not connected to power) on minimum performance mode using a lifecam the code now runs at around 30 fps almost never dropping bellow 27 fps. however due to how tag detection works when you're NOT viewing a tag the fps can drop significantly to around 17 fps, however we really don't need it to be any higher when were not viewing a tag.

i did a very rough estimation comparing the specs of my computer to the jetson and so and ive concluded that the jetson should probably run the code at around 20 fps when viewing a tag (using a lifecam). which is acceptable.